# swagger.yaml
swagger: "2.0"
info:
  description: "URL shortener written in Go"
  version: "1.0.0"
  title: "URL shortener"
host: "localhost"
basePath: "/"
...
schemes:
  - "http"
paths:
...
  /api/count/{urlKey}:
    get:
      tags:
        - "URL shortener"
      summary: "Redirections count"
      description: "Return the redirections count for the specified url key. It may not work if the shortned URL is http only (no https)"
      produces:
        - "application/json"
      parameters:
        - name: "urlKey"
          in: "path"
          description: "Short URL key used to delete the original website from the DB."
          required: true
          type: "string"
      responses:
        "200":
          description: "Return the redirection count"
          schema:
            $ref: "#/definitions/RedirectionsCountResponse"
        "404":
          description: "URL key not found"
        "500":
          description: "Internal Server Error"
definitions:
...
  RedirectionsCountResponse:
    description: "Contains the shorten URL and the its key"
    type: "object"
    properties:
      redirectionsCount:
        description: "How many time the short url has been used for redirecting to the original URL"
        type: "integer"
        example: 1